Use intptr_t for intermediate cast to gnutls_transport_ptr_t
In Win64, 'long' is still only 32 bits, and isn't enough to avoid the
"cast to pointer from integer of different size" warning. This kind of
braindamage is why intptr_t exists, so use it.

+++ dtls.c
@@ -363,7 +363,7 @@ static int start_dtls_handshake(struct openconnect_info *vpninfo, int dtls_fd)
 	}
 
 	gnutls_transport_set_ptr(dtls_ssl,
-				 (gnutls_transport_ptr_t)(long) dtls_fd);
+				 (gnutls_transport_ptr_t)(intptr_t)dtls_fd);
 
 	gnutls_record_disable_padding(dtls_ssl);
 	master_secret.data = vpninfo->dtls_secret;
